Rewrote functionality to send queries to RU using template files only, discarding trying to use xml related functionality. This approach is much more robust.

+++ channelmanagement_rentalsunited_dictionaries.class.php
@@ -27,10 +27,6 @@ class channelmanagement_rentalsunited_dictionaries
 			}
 		}
 		
-		$JRUser									= jomres_singleton_abstract::getInstance( 'jr_user' );
-		
-		
-		
 		$this->dictionaries = array();
 		
 		$siteConfig = jomres_singleton_abstract::getInstance('jomres_config_site_singleton');
@@ -44,27 +40,39 @@ class channelmanagement_rentalsunited_dictionaries
 			throw new Exception( jr_gettext('CHANNELMANAGEMENT_RENTALSUNITED_PASSWORD_NOT_SET','CHANNELMANAGEMENT_RENTALSUNITED_PASSWORD_NOT_SET',false) );
 		}
 		
-		jr_import('channelmanagement_rentalsunited_communication');
-		$this->channelmanagement_rentalsunited_communication = new channelmanagement_rentalsunited_communication();
-		$this->channelmanagement_rentalsunited_communication->set_username($jrConfig['channel_manager_framework_user_accounts']['rentalsunited']["channel_management_rentals_united_username"]);
-		$this->channelmanagement_rentalsunited_communication->set_password($jrConfig['channel_manager_framework_user_accounts']['rentalsunited']["channel_management_rentals_united_password"]);
-		
 		$dictionary_array = $this->get_dictionary_initialisation_array();
 
-		
+
+        set_showtime("property_managers_id" , "system");
+
 		if (!empty($dictionary_array)) {
+            jr_import('channelmanagement_rentalsunited_communication');
+            $channelmanagement_rentalsunited_communication = new channelmanagement_rentalsunited_communication();
+
 			foreach ($dictionary_array as $dictionary) {
+
 				if (!file_exists(JOMRES_CHANNEL_DICTIONARIES.JRDS.$this->current_channel_manager.JRDS.$dictionary.'.json')) {
-					$response = $this->channelmanagement_rentalsunited_communication->communicate( array() , $dictionary );
-					if ($response != false ) {
-						$encoded = json_encode($response);
-						$file_contents = str_replace("@attributes" , "xml_attributes" ,$encoded );
-						file_put_contents( JOMRES_CHANNEL_DICTIONARIES.JRDS.$this->current_channel_manager.JRDS.$dictionary.'.json' , $file_contents );
-					} else {
-						file_put_contents( JOMRES_CHANNEL_DICTIONARIES.JRDS.$this->current_channel_manager.JRDS.$dictionary.'.json' , '' );
-						logging::log_message("Could not receive data from remote service, empty content returned for dictionary ".$dictionary, 'channel_management', 'ERROR' , $this->current_channel_manager );
-					}
-					
+                    if (file_exists(RENTALS_UNITED_PLUGIN_ROOT.'templates'.JRDS."xml".JRDS.$dictionary.'.xml')) {
+                        $auth = get_auth();
+
+                        $output = array("AUTHENTICATION" => $auth);
+
+                        $tmpl = new patTemplate();
+                        $tmpl->addRows('pageoutput', array($output));
+                        $tmpl->setRoot(RENTALS_UNITED_PLUGIN_ROOT . 'templates' . JRDS . "xml");
+                        $tmpl->readTemplatesFromInput($dictionary . '.xml');
+                        $xml_str = $tmpl->getParsedTemplate();
+
+                        $response = $channelmanagement_rentalsunited_communication->communicate($dictionary, $xml_str);
+                        if ($response != false) {
+                            $encoded = json_encode($response, JSON_PRETTY_PRINT);
+                            $file_contents = str_replace("@attributes", "xml_attributes", $encoded);
+                            file_put_contents(JOMRES_CHANNEL_DICTIONARIES . JRDS . $this->current_channel_manager . JRDS . $dictionary . '.json', $file_contents);
+                        } else {
+                            file_put_contents(JOMRES_CHANNEL_DICTIONARIES . JRDS . $this->current_channel_manager . JRDS . $dictionary . '.json', '');
+                            logging::log_message("Could not receive data from remote service, empty content returned for dictionary " . $dictionary, 'channel_management', 'ERROR', $this->current_channel_manager);
+                        }
+                    }
 				} 
 			}
